fixed win percentage

+++ src/main/kotlin/org/allenterescenco/tournamentmanager/console/models/team/TeamModel.kt
@@ -4,6 +4,6 @@ data class TeamModel(var id: Long = 0,
                      var name: String = "Default Name",
                      var wins: Int = 0,
                      var losses: Int = 0,
-                     var winPercentage: Double = 0.00,
+                     var winPercentage: Int = 0,
                      var players: Array<String> = arrayOf("",""),
 )
